+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+* New Spinner component ([#57](https://github.com/narqo/react-islands/pull/57)).
+
 ## 0.3.1 - 2016-06-05
 
 ### Fixed